+++ plugins/repository-azure/src/main/java/org/opensearch/repositories/azure/AzureBlobContainer.java
@@ -40,6 +40,7 @@ import org.apache.logging.log4j.Logger;
 import org.opensearch.action.ActionRunnable;
 import org.opensearch.action.support.GroupedActionListener;
 import org.opensearch.action.support.PlainActionFuture;
+import org.opensearch.cluster.metadata.CryptoMetadata;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.blobstore.BlobContainer;
 import org.opensearch.common.blobstore.BlobMetadata;
@@ -134,6 +135,18 @@ public class AzureBlobContainer extends AbstractBlobContainer {
         } catch (URISyntaxException | BlobStorageException e) {
             throw new IOException("Can not write blob " + blobName, e);
         }
+    }
+
+    @Override
+    public void writeBlobWithMetadata(
+        String blobName,
+        InputStream inputStream,
+        long blobSize,
+        boolean failIfAlreadyExists,
+        @Nullable Map<String, String> metadata,
+        @Nullable CryptoMetadata cryptoMetadata
+    ) throws IOException {
+        writeBlob(blobName, inputStream, blobSize, failIfAlreadyExists);
     }
 
     @Override