+++ server/src/main/java/org/opensearch/snapshots/SnapshotShardsService.java
@@ -470,6 +470,9 @@ public class SnapshotShardsService extends AbstractLifecycleComponent implements
                     // we flush first to make sure we get the latest writes snapshotted
                     wrappedSnapshot = indexShard.acquireLastIndexCommit(true);
                     final IndexCommit snapshotIndexCommit = wrappedSnapshot.get();
+
+                    IndexMetadata indexMetadata = clusterService.state().metadata().index(indexId.getName());
+
                     repository.snapshotShard(
                         indexShard.store(),
                         indexShard.mapperService(),
@@ -480,7 +483,8 @@ public class SnapshotShardsService extends AbstractLifecycleComponent implements
                         snapshotStatus,
                         version,
                         userMetadata,
-                        ActionListener.runBefore(listener, wrappedSnapshot::close)
+                        ActionListener.runBefore(listener, wrappedSnapshot::close),
+                        indexMetadata
                     );
                 }
             } catch (Exception e) {